requirements.txt
flask
mysql-connector-python
requests